src/1.getting_started/1.2.hello_window_clear/requirements.txt
PyOpenGL==3.1.5
glfw==2.1.0